change TableChangeSet to StateChangeSet
