+++ crates/rooch-server/src/jsonrpc_types/mod.rs
@@ -8,7 +8,7 @@ mod execute_tx_response;
 mod function_return_value_view;
 mod move_types;
 mod rooch_types;
-mod table_view;
+mod state_view;
 mod transaction_argument_view;
 
 pub mod bytes;
@@ -17,6 +17,6 @@ pub use self::rooch_types::*;
 pub use execute_tx_response::*;
 pub use function_return_value_view::*;
 pub use move_types::*;
+pub use state_view::*;
 pub use str_view::*;
-pub use table_view::*;
 pub use transaction_argument_view::*;
