+++ moveos/moveos-types/src/lib.rs
@@ -14,6 +14,5 @@ pub mod move_types;
 pub mod object;
 pub mod state;
 pub mod storage_context;
-pub mod table;
 pub mod transaction;
 pub mod tx_context;
