+++ moveos/moveos-types/src/state.rs
@@ -1,12 +1,15 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+use std::collections::{BTreeMap, BTreeSet};
+
 use crate::{
     addresses::MOVEOS_STD_ADDRESS,
-    object::{self, AnnotatedObject, Object, RawObject},
+    object::{self, AnnotatedObject, Object, ObjectID, RawObject},
 };
 use anyhow::{bail, ensure, Result};
 use move_core_types::{
+    effects::Op,
     language_storage::{StructTag, TypeTag},
     move_resource::MoveStructType,
     resolver::MoveResolver,
@@ -165,4 +168,45 @@ impl AnnotatedState {
             _ => bail!("Expect MoveStruct but found {:?}", self.move_value),
         }
     }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct TableTypeInfo {
+    pub key_type: TypeTag,
+}
+
+impl TableTypeInfo {
+    pub fn new(key_type: TypeTag) -> Self {
+        Self { key_type }
+    }
+}
+
+impl std::fmt::Display for TableTypeInfo {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "Table<{}>", self.key_type)
+    }
+}
+
+/// Global State change set.
+#[derive(Default, Clone, Debug)]
+pub struct StateChangeSet {
+    pub new_tables: BTreeMap<ObjectID, TableTypeInfo>,
+    pub removed_tables: BTreeSet<ObjectID>,
+    pub changes: BTreeMap<ObjectID, StateChange>,
+}
+
+/// A change of a single state.
+#[derive(Clone, Debug)]
+pub struct StateChange {
+    //TODO should we keep the key's type here?
+    pub entries: BTreeMap<Vec<u8>, Op<State>>,
+}
+
+/// A global state resolver which needs to be provided by the environment.
+/// This allows to lookup data in remote storage.
+/// If the handle is GLOBAL_OBJECT_STORAGE_HANDLE, it will get the data from the global state tree,
+/// otherwise it will get the data from the table state tree.
+/// The key can be an ObjectID or an arbitrary key of a table.
+pub trait StateResolver {
+    fn resolve_state(&self, handle: &ObjectID, key: &[u8]) -> Result<Option<State>, anyhow::Error>;
 }
