+++ moveos/moveos/src/vm/mod.rs
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use move_core_types::resolver::MoveResolver;
-use moveos_types::table::TableResolver;
+use moveos_types::state::StateResolver;
 
 pub mod dependency_order;
 pub mod move_vm_ext;
 pub mod tx_argument_resolver;
 pub mod vm_status_explainer;
 
-pub trait MoveResolverExt: MoveResolver + TableResolver {}
+pub trait MoveResolverExt: MoveResolver + StateResolver {}
 
-impl<T> MoveResolverExt for T where T: MoveResolver + TableResolver {}
+impl<T> MoveResolverExt for T where T: MoveResolver + StateResolver {}
